// app/types.ts
export interface PostProps {
  slug: string;
  title: string;
  description: string;
  content: string;
  live?: string;
  code?: string;
  image?: string;
  image2?: string;
  image3?: string;
  image4?: string;
  image5?: string;
  image6?: string;
  image7?: string;
  image8?: string;
  image9?: string;
  date: string;
  codeText: string;
  liveText: string;
}